requirements.txt
piazza-api
anthropic
discord-webhook
python-dotenv
chromadb
sentence-transformers
pypdf
